requirements.txt
astroid==2.3.3
certifi==2019.9.11
Django==2.2.7
django-chartit==0.2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.17.4
pandas==0.25.3
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
tqdm==4.39.0
wrapt==1.11.2
